refactor: 컴포넌트 폴더 구조를 colocation pattern 형식으로 수정
