+++ src/components/common/Select/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./Select";
+export type { SelectProps } from "./Select";
