+++ src/components/feature/CardInfoFormSection/components/CardValidityPeriodInputField/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CardValidityPeriodInputField";
