+++ src/stories/Button.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 
-import Button from "../components/common/Button.tsx";
+import Button from "../components/common/Button";
 
 const meta = {
   title: "Button",
