+++ src/stories/CardValidityPeriodInputField.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 
-import useFormWrapper from "../components/common/FormContainer";
-import CardValidityPeriodInputField from "../components/feature/CardInfoFormSection/components/CardValidityPeriodInputField/CardValidityPeriodInputField";
+import useFormWrapper from "@hooks/useFormWrapper";
+import CardValidityPeriodInputField from "../components/feature/CardInfoFormSection/components/CardValidityPeriodInputField";
 import { INITIAL_CARD_INFO_FORM_STATE } from "../components/feature/CardInfoFormSection/formState";
 
 const meta = {
